prettier

+++ src/test/DepositAndRefund.t.sol
@@ -6,22 +6,27 @@ import "../DepositAndRefund.sol";
 
 interface CheatCodes {
     function assume(bool) external;
+
     function prank(address) external;
+
     function deal(address who, uint256 newBalance) external;
+
     function expectRevert(bytes calldata) external;
+
     function warp(uint256) external;
 }
 
 contract DepositAndRefundTest is DSTest {
-    DepositAndRefund dnr; 
+    DepositAndRefund dnr;
     CheatCodes cheats = CheatCodes(HEVM_ADDRESS);
 
     receive() external payable {}
+
     function setUp() public {
         dnr = new DepositAndRefund();
     }
 
-    function testDeposit(address _party, uint _amount) public {
+    function testDeposit(address _party, uint256 _amount) public {
         cheats.assume(_amount > 0);
         cheats.deal(_party, _amount);
         cheats.prank(_party);
@@ -29,7 +34,10 @@ contract DepositAndRefundTest is DSTest {
         assert(dnr.getBalance(_party) == _amount);
     }
 
-    function testInsufficientBalance(uint _amountToDeposit, uint _amountToWithdraw) public {
+    function testInsufficientBalance(
+        uint256 _amountToDeposit,
+        uint256 _amountToWithdraw
+    ) public {
         cheats.assume(_amountToDeposit > 0);
         cheats.assume(_amountToWithdraw > _amountToDeposit);
         cheats.deal(address(this), _amountToDeposit);
@@ -38,7 +46,11 @@ contract DepositAndRefundTest is DSTest {
         dnr.withdraw(_amountToWithdraw);
     }
 
-    function testWithdrawPostLockup(uint _amountToDeposit, uint _amountToWithdraw, uint _elapsedTime) public {
+    function testWithdrawPostLockup(
+        uint256 _amountToDeposit,
+        uint256 _amountToWithdraw,
+        uint256 _elapsedTime
+    ) public {
         cheats.assume(_amountToDeposit > 0);
         cheats.assume(_amountToWithdraw <= _amountToDeposit);
         cheats.assume(_elapsedTime >= dnr.lockupPeriod());
@@ -48,7 +60,11 @@ contract DepositAndRefundTest is DSTest {
         dnr.withdraw(_amountToWithdraw);
     }
 
-    function testWithdrawPreLockup(uint _amountToDeposit, uint _amountToWithdraw, uint _elapsedTime) public {
+    function testWithdrawPreLockup(
+        uint256 _amountToDeposit,
+        uint256 _amountToWithdraw,
+        uint256 _elapsedTime
+    ) public {
         cheats.assume(_amountToDeposit > 0);
         cheats.assume(_amountToWithdraw <= _amountToDeposit);
         cheats.assume(_elapsedTime < dnr.lockupPeriod());
